move xchain transfer into separate file
