+++ slk/sidechain.py
@@ -18,80 +18,15 @@ import time
 from multiprocessing import Process, Value
 from typing import Callable, List
 
-from xrpl.models import Amount, Memo, Payment
-
 import slk.interactive as interactive
 from slk.chain import Chain, configs_for_testnet, single_node_chain
 from slk.chain_setup import setup_mainchain, setup_sidechain
-from slk.common import Account, disable_eprint, eprint
+from slk.common import disable_eprint, eprint
 from slk.config_file import ConfigFile
 from slk.log_analyzer import convert_log
 from slk.sidechain_params import SidechainParams
 from slk.testnet import sidechain_network
-
-
-def _xchain_transfer(
-    from_chain: Chain,
-    to_chain: Chain,
-    src: Account,
-    dst: Account,
-    amt: Amount,
-    from_chain_door: Account,
-    to_chain_door: Account,
-):
-    memo = Memo(memo_data=dst.account_id_str_as_hex())
-    from_chain(
-        Payment(
-            account=src.account_id,
-            destination=from_chain_door.account_id,
-            amount=amt,
-            memos=[memo],
-        )
-    )
-    from_chain.maybe_ledger_accept()
-    if to_chain.standalone:
-        # from_chain (side chain) sends a txn, but won't close the to_chain (main chain)
-        # ledger
-        time.sleep(1)
-        to_chain.maybe_ledger_accept()
-
-
-def main_to_side_transfer(
-    mc_chain: Chain,
-    sc_chain: Chain,
-    src: Account,
-    dst: Account,
-    amt: Amount,
-    params: SidechainParams,
-):
-    _xchain_transfer(
-        mc_chain,
-        sc_chain,
-        src,
-        dst,
-        amt,
-        params.mc_door_account,
-        params.sc_door_account,
-    )
-
-
-def side_to_main_transfer(
-    mc_chain: Chain,
-    sc_chain: Chain,
-    src: Account,
-    dst: Account,
-    amt: Amount,
-    params: SidechainParams,
-):
-    _xchain_transfer(
-        sc_chain,
-        mc_chain,
-        src,
-        dst,
-        amt,
-        params.sc_door_account,
-        params.mc_door_account,
-    )
+from slk.xchain_transfer import main_to_side_transfer, side_to_main_transfer
 
 
 def simple_test(mc_chain: Chain, sc_chain: Chain, params: SidechainParams):
